+++ cl/models/gpt.py
@@ -1,0 +1,2 @@
+class FastGPT2DistilModel:
+    pass
