+++ cl/utils/data_utils.py
@@ -1,10 +1,11 @@
 from itertools import chain
 
+
 class DataPreprocessing:
     def __init__(self, block_size, metric):
         self.block_size = block_size
         self.metric = metric
-    
+
     def group_texts(self, examples):
         concatenator = lambda x: list(chain(*x)) if isinstance(x[0], list) else x
         concatenated_examples = {k: concatenator(examples[k]) for k in examples.keys()}
@@ -12,7 +13,10 @@ class DataPreprocessing:
         if total_length >= self.block_size:
             total_length = (total_length // self.block_size) * self.block_size
         result = {
-            k: [t[i : i + self.block_size] for i in range(0, total_length, self.block_size)]
+            k: [
+                t[i : i + self.block_size]
+                for i in range(0, total_length, self.block_size)
+            ]
             for k, t in concatenated_examples.items()
         }
         result["labels"] = result["input_ids"].copy()
